resume_mvp/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
email-validator==2.2.0
fpdf2==2.8.2
